% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighestNumber.R
\name{vector_max}
\alias{vector_max}
\title{Finding the highest number in a vector}
\usage{
vector_max(vector)
}
\description{
This function finds the biggest number in a vector
}
\examples{
vector_max()
}

